Change the way room password is treated
In general, before this commit room password was saved as class member and
stored as cleartext til the program would end. Now, I store the encrypted key
and salt for it, along with the randomly generated key for it. I send the
enctypted message, it's salt, encrypted key and it's hash in every message.
Don't know if it's a good idea, but I'll leave it like this until I get a better
solution.
I also added anti-recursive-clipboard-update.